+++ src/main/java/insane96mcp/enhancedai/modules/base/feature/Targeting.java
@@ -46,7 +46,7 @@ public class Targeting extends Feature {
 	public static MinMax followRangeOverride = new MinMax(24, 48);
 	@Config(min = 0d, max = 128d)
 	@Label(name = "XRay Range Override", description = "How far away can the mobs see the player even through walls. Setting 'Max' to 0 will make mobs not able to see through walls. I recommend using mods like Mobs Properties Randomness to have more control over the attribute; the attribute name is 'enhancedai:generic.xray_follow_range'.")
-	public static MinMax xrayRangeOverride = new MinMax(16, 32);
+	public static MinMax xrayRangeOverride = new MinMax(12, 24);
 	@Config
 	@Label(name = "Instant Target", description = "Mobs will no longer take random time to target a player.")
 	public static Boolean instaTarget = true;