+++ changelog.md
@@ -6,6 +6,7 @@
 * Animals now scare only animals of the same species
 * Reduced follow range override (32~64 -> 24~48)
 * Reduced xray distance (16~32 -> 12~24)
+* Reduced Throwing Web Spiders range and slowness applied
 
 ## 1.9.1
 * Only 20% of animals are now able to fight back, the others will just scatter